Extract arrow_helpers module
